+++ LookupAnything/Properties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("LookupAnything")]
-[assembly: AssemblyVersion("1.20.1")]
-[assembly: AssemblyFileVersion("1.20.1")]
+[assembly: AssemblyVersion("1.21.0")]
+[assembly: AssemblyFileVersion("1.21.0")]