+++ TractorMod/Properties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("TractorMod")]
-[assembly: AssemblyVersion("4.7.2")]
-[assembly: AssemblyFileVersion("4.7.2")]
+[assembly: AssemblyVersion("4.7.3")]
+[assembly: AssemblyFileVersion("4.7.3")]